requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.22
alembic==1.11.1
python-dotenv==1.0.0
pydantic==2.5.1
passlib[argon2]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
python-jose[cryptography]==3.3.0